# Web Platform/robomap_backend/public/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: MIM
  description: 'Collection of all api requests'
  contact:
    email: support@mim.com
  version: 0.0.1
servers:
  -
    url: 'http://localhost:8000/api'
paths:
...
  /roles:
    get:
      description: 'Returns a list with all registered roles that meet the parameters requirements'
      operationId: 'get roles'
      responses:
        '200':
          description: 'Request successful'
    put:
      description: 'Edits the role with the specified id'
      operationId: 'edit role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                id:
                  type: number
                  default: 21
                name:
                  type: string
                  default: 'New Role21'
              type: object
      responses:
        '200':
          description: 'Request successful'
        '404':
          description: 'Role not found'
    post:
      operationId: 'create role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  default: Role21
              type: object
      responses:
        '400':
          description: 'Wrong parameters'
        '200':
          description: 'Request successful'
          content:
            application/json:
              schema:
                properties:
                  id: { type: number, default: 21 }
                  name: { type: string, default: Role21 }
                type: object
  '/roles/{id}':
    get:
      description: 'Returns the specified role'
      operationId: 'get role'
      parameters:
        -
          name: id
          in: path
          description: 'Role id'
          required: true
          schema:
            type: number
            default: 21
      responses:
        '200':
          description: 'Request successful'
          content:
            application/json:
              schema:
                properties:
                  id: { type: number, default: 21 }
                  name: { type: string, default: Role21 }
                type: object
        '404':
          description: 'Role not found'
    delete:
      description: 'Deletes the role with the specified id'
      operationId: 'delete role'
      parameters:
        -
          name: id
          in: path
          description: 'Role id'
          required: true
          schema:
            type: number
            default: 21
      responses:
        '200':
          description: 'Request successful'
        '404':
          description: 'Role not found'
...